+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## v0.12.0 (2023-11-01)
+
 - Add `:client` and `:region` options to stack and stack-properties
   components.
 - Move `anomaly?`, `aws-error-code`, and `aws-message-fns` to `salmon.util`